Added Wills Obsessive Packing operator.
Output (in FUN file) is wrong. Every objective value gets the same answer and it appears to not be randomised at all. I have no idea why and no time to figure it out right now.
Maybe one of you could figure it out - I'll talk to Markus on Monday as well if no-one has figured it out by then.